% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_pathview_gen.R
\name{get_all_fold_changes}
\alias{get_all_fold_changes}
\title{get_all_fold_changes}
\usage{
get_all_fold_changes(lst.deg.res, species, pval.cutoff = 0.01)
}
\arguments{
\item{lst.deg.res}{DEG results from SCODA (a named list of data frames)}

\item{species}{'human' or 'mouse'.}

\item{pval.cutoff}{p-value cutoff to filter DEGs.}
}
\value{
a named list of named list of fold changes. Names of fold changes are ENTREZ gene code.
}
\description{
This function retreives fold changes from the SCODA DEG results.
}
