% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_pathview_gen.R
\name{save_kegg_pathviews}
\alias{save_kegg_pathviews}
\title{save_kegg_pathviews}
\usage{
save_kegg_pathviews(
  adata,
  target_cell,
  df_pathways_map,
  pathways = NULL,aC8`C8
  deg.p.val.cutoff = 0.01,
  gsa.p.val.cutoff = 0.01,
  deg.key = "DEG_vs_ref",
  gsa.key = "GSA_vs_ref_up"
)
}
\arguments{
\item{adata}{SCODA-processed anndata object.}

\item{target_cell}{target cell-type to retreives its DEGs on the KEGG pathview images. Should be one of the name list of lst.gsa.all.}

\item{df_pathways_map}{A data frame containing the mapping between the KEGG pathway and the pathway used in the SCODA. Use the output of the function 'get_pathways_map'.}

\item{deg.p.val.cutoff}{p-value to filter the DEG analysis results in the SCODA result file.}

\item{gsa.p.val.cutoff}{p-value to filter the Gene Set (Enrichment) analysis results in the SCODA result file.}
}
\value{
The name of the folder where KEGG pathview images are stored. It is 'KEGG_pathview_(cell type)', where (cell type) is the 'target_cell'.
}
\description{
This function saves Kthe EGG pathview images colored by log-fold-changes of the pathways identified in the SCODA gene set (enrichment) analysis result for a specific cell-type.
}
